clean up and move syncing code into separate files
